+++ src/RemoteOps.Desktop/Views/LogsView.xaml.cs
@@ -1,0 +1,11 @@
+using System.Windows.Controls;
+
+namespace RemoteOps.Desktop.Views;
+
+public partial class LogsView : UserControl
+{
+    public LogsView()
+    {
+        InitializeComponent();
+    }
+}